Added pundit task policy

+++ app/controllers/tasks_controller.rb
@@ -1,17 +1,19 @@
 # frozen_string_literal: true
 
 class TasksController < ApplicationController
-  before_action :authenticate_user_using_x_auth_token, except: [:new, :edit]
-  before_action :load_task, only: %i[show update destroy]
   before_action :authenticate_user_using_x_auth_token
+  before_action :load_task, only: %i[show update destroy]
+  after_action :verify_authorized, except: :index
+  after_action :verify_policy_scoped, only: :index
 
   def index
-    tasks = Task.all
+    tasks = policy_scope(Task)
     render status: :ok, json: { tasks: tasks }
-   end
+  end
 
   def create
     task = Task.new(task_params.merge(task_owner_id: @current_user.id))
+    authorize task
     if task.save
       render status: :ok, json: { notice: t("successfully_created", entity: "Task") }
     else
@@ -21,12 +23,13 @@ class TasksController < ApplicationController
   end
 
   def show
-    render
+    authorize @task
   end
 
   def update
+    authorize @task
     if @task.update(task_params)
-      render status: :ok, json: { notice: "Successfully updated task." }
+      render status: :ok, json: {}
     else
       render status: :unprocessable_entity,
         json: { error: @task.errors.full_messages.to_sentence }
@@ -34,8 +37,9 @@ class TasksController < ApplicationController
   end
 
   def destroy
+    authorize @task
     if @task.destroy
-      render status: :ok, json: { notice: "Successfully deleted task." }
+      render status: :ok, json: {}
     else
       render status: :unprocessable_entity,
         json: { error: @task.errors.full_messages.to_sentence }
@@ -44,14 +48,14 @@ class TasksController < ApplicationController
 
   private
 
+    def task_params
+      params.require(:task).permit(:title, :assigned_user_id)
+    end
+
     def load_task
       @task = Task.find_by(slug: params[:slug])
       unless @task
         render status: :not_found, json: { error: t("task.not_found") }
       end
-  end
-
-    def task_params
-      params.require(:task).permit(:title, :user_id)
     end
 end
